requirements.txt
pyTelegramBotAPI
docx2pdf
img2pdf
pytesseract
Pillow
python-dotenv